#27343 beta for outsiders

https://ticket.office.sevenval.de/issues/27343

See merge request !5

+++ setup/switch-version-internal.sh
@@ -1,11 +1,23 @@
 
-set -e  
+VERSION=$1
+CHANNEL=fit14${VERSION#14}
+
+#27343 read credentials for beta users
+CREDENTIALS=$(cat /vagrant/credentials 2>/dev/null)
+if [ -n "$CREDENTIALS" ]; then
+	CREDENTIALS="${CREDENTIALS}@"
+fi
+
+if ! curl -m 5 -s -f "https://${CREDENTIALS}download.sevenval-fit.com/FIT_Server_Beta/14/check-credentials?version=${VERSION}" >/dev/null; then
+	echo -e "Sorry, you don't have credentials to switch builds.\n\nConfigure your credentials like this:\n\$ echo 'user:pass' > ~/fit14-devbox/credentials\n\nContact Sevenval if you don't have download server access, yet."
+	exit 126
+fi
+
+set -e
 
 trap 'echo "ERROR: could not install new FIT packages. try switching back to stable by running
 	switch-version.sh stable"' SIGINT SIGTERM
 
-VERSION=$1
-CHANNEL=fit14${VERSION#14}
 
 echo "switching FIT installation to $VERSION"
 REMOVE="`rpm -qa | grep ^fit14`" || true
@@ -22,7 +34,7 @@ if ! yum-config-manager --disablerepo=* --enablerepo=${CHANNEL} 2>&1 >/dev/null;
 
 	echo "[${CHANNEL}]
 name = Sevenval FIT Server 14 Channel ${CHANNEL}
-baseurl = https://download.sevenval-fit.com/FIT_Server_Beta/14/packages/RHEL/7/${VERSION}/x86_64
+baseurl = https://${CREDENTIALS}download.sevenval-fit.com/FIT_Server_Beta/14/packages/RHEL/7/${VERSION}/x86_64
 enabled = 1
 gpgkey = https://download.sevenval-fit.com/FIT_Server/sevenval.key
 gpgcheck = 0
@@ -32,6 +44,8 @@ fi
 yum-config-manager --quiet --enable ${CHANNEL} >/dev/null
 yum --disableplugin=fastestmirror install -y fit14fitserver
 
+#27343 don't store credentials inside the box
+sed -i '/baseurl.*@/ s/https:\/\/.*\?@/https:\/\//' /etc/yum.repos.d/fit14-tmp.repo
 
 cd /opt/sevenval/fit14/conf
